+++ clang-tidy/nodecpp/NodeCppTidyModule.cpp
@@ -13,6 +13,7 @@
 #include "ArrayTypeCheck.h"
 #include "NakedPtrAssignmentCheck.h"
 #include "NakedPtrFieldCheck.h"
+#include "NakedPtrFromMethodCheck.h"
 #include "NakedPtrFromReturnCheck.h"
 #include "NakedPtrFuncCheck.h"
 #include "NewExprCheck.h"
@@ -34,6 +35,8 @@ public:
         "nodecpp-naked-ptr-assignment");
     CheckFactories.registerCheck<NakedPtrFieldCheck>(
         "nodecpp-naked-ptr-field");
+    CheckFactories.registerCheck<NakedPtrFromMethodCheck>(
+        "nodecpp-naked-ptr-from-method");
     CheckFactories.registerCheck<NakedPtrFromReturnCheck>(
         "nodecpp-naked-ptr-from-return");
     CheckFactories.registerCheck<NakedPtrFuncCheck>(